Sort out components into more decoupled pieces
